+++ hub/src/mixer/obs/ObsConfiguration.spec.ts
@@ -59,11 +59,13 @@ describe('fromJson/toJson', () => {
         const conf = createDefaultObsConfiguration()
         conf.setIp("1.2.3.4")
         conf.setPort(1234)
+        conf.setLiveMode("stream")
         const loadedConf = createDefaultObsConfiguration()
         loadedConf.fromJson(conf.toJson())
         
         expect(loadedConf.getIp().toString()).toEqual("1.2.3.4")
         expect(loadedConf.getPort().toNumber()).toEqual(1234)
+        expect(loadedConf.getLiveMode()).toEqual("stream")
     })
 })
 
@@ -72,10 +74,12 @@ describe('clone', () => {
         const conf = createDefaultObsConfiguration()
         conf.setIp("1.2.3.4")
         conf.setPort(1234)
+        conf.setLiveMode("stream")
         const clone = conf.clone()
         conf.setIp("2.3.4.5") // it should be a new instance
         
         expect(clone.getIp().toString()).toEqual("1.2.3.4")
         expect(clone.getPort().toNumber()).toEqual(1234)
+        expect(clone.getLiveMode()).toEqual("stream")
     })
 })