+++ hub/src/mixer/obs/ObsConfiguration.ts
@@ -5,26 +5,34 @@ import {Configuration} from '../interfaces'
 export type ObsConfigurationSaveType = {
     ip: string
     port: number
+    liveMode?: ObsConfigurationLiveMode
 }
+
+export type ObsConfigurationLiveMode = "always" | "stream" | "record" | "streamOrRecord"
 
 class ObsConfiguration extends Configuration {
     ip: IpAddress
     port: IpPort
+    liveMode: ObsConfigurationLiveMode
+
 
     constructor() {
         super()
         this.ip = ObsConfiguration.defaultIp
         this.port = ObsConfiguration.defaultPort
+        this.liveMode = ObsConfiguration.defaultLiveMode
     }
 
     fromJson(data: ObsConfigurationSaveType): void {
         this.loadIpAddress("ip", this.setIp.bind(this), data)
         this.loadIpPort("port", this.setPort.bind(this), data)
+        data.liveMode && this.setLiveMode(data.liveMode)
     }
     toJson(): ObsConfigurationSaveType {
         return {
             ip: this.ip.toString(),
             port: this.port.toNumber(),
+            liveMode: this.liveMode,
         }
     }
     clone(): ObsConfiguration {
@@ -61,8 +69,21 @@ class ObsConfiguration extends Configuration {
         return this.port
     }
 
+    setLiveMode(mode: ObsConfigurationLiveMode) {
+        this.liveMode = mode
+    }
+    getLiveMode() {
+        return this.liveMode
+    }
+
     private static readonly defaultIp = ipAddress("127.0.0.1")
     private static readonly defaultPort = ipPort(4444)
+    private static readonly defaultLiveMode = "always"
+
+    static isValidLiveMode(theString: string): theString is ObsConfigurationLiveMode {
+        return theString === "always" || theString === "streamOrRecord" || theString === "stream" || theString === "record"
+    }
+
 }
 
 export default ObsConfiguration